+++ src/App.css
@@ -36,3 +36,18 @@
     transform: rotate(360deg);
   }
 }
+
+.container {
+  display: flex;
+  flex-direction: row;
+}
+
+.appContainer {
+  height: 100%;
+  width: 80%;
+}
+
+.pages {
+  height: 80%;
+  width: 100%;
+}
