+++ src/App.js
@@ -5,17 +5,21 @@ import Login from "./pages/Login/Login";
 import Home from "./pages/Home/Home";
 import LandPage from "./pages/LandPage/LandPage";
 import Cadastrar from "./pages/Cadastrar/Cadastrar";
-import ListaDeTarefas from "./pages/Home/ListadeTarefas";
-import ListaDeProjetos from "./pages/Home/ListaDeProjetos";
+// import ListaDeTarefas from "./pages/Home/ListadeTarefas";
+// import ListaDeProjetos from "./pages/Home/ListaDeProjetos";
 import Calendario from "./pages/Calendario/Calendario";
-import Projetos from "./pages/Projetos/Projetos"
-import Config from "./pages/Config/Config"
+import Projetos from "./pages/Projetos/Projetos";
+import Config from "./pages/Config/Config";
+import "@fontsource/montserrat";
+import Sidebar from "./Componets/Sidebar/Sidebar";
+import "./App.css";
+import Header from "./Componets/Header/Header";
 
 function App() {
   const router = createBrowserRouter([
     {
       path: "/login",
-      element: <Login />,
+      element: <Login/>,
     },
     {
       path: "/",
@@ -27,27 +31,35 @@ function App() {
     },
     {
       path: "/home",
-      element: <ListaDeTarefas/>, 
+      element: <Home />,
     },
     {
       path: "/calendario",
-      element: <Calendario/>
+      element: <Calendario />,
     },
     {
       path: "/projetos",
-      element: <Projetos/>
+      element: <Projetos />,
     },
     {
       path: "/config",
-      element: <Config/>
+      element: <Config />,
     },
   ]);
 
   return (
-    <>
+    
+    <div className="container">
+      <Sidebar />
+      <div className="appContainer">
+        <Header />
+        <div className="pages">
+          <RouterProvider router={router} />
+        </div>
+      </div>
       <ToastContainer />
-      <RouterProvider router={router} />
-    </>
+    </div>
+    
   );
 }
