+++ src/Componets/Header/Header.module.css
@@ -1,0 +1,18 @@
+.container {
+  background-color: #FFFFFF;
+  width: 100%;
+  height: 15%;
+  justify-content: right;
+  text-align: right;
+  box-shadow: rgba(0, 0, 0, 0.15) 2.4px 2.4px 3.2px;
+}
+
+
+button{
+  height: 50px;
+  width: 135px;
+  border-radius: 15px;
+  background-color: #222831;
+  color: white;
+  margin: 10px;
+}
