+++ src/Componets/Sidebar/Sidebar.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import styles from "./Sidebar.module.css";
+import { Link } from "react-router-dom";
+
+export default function Sidebar() {
+  return (
+    <div className={styles.container}>
+
+    </div>
+  );
+}
