+++ src/pages/Cadastrar/Cadastrar.css
@@ -4,8 +4,8 @@
   }
   
   * {
-    font-family: "Inter", sans-serif;
-    font-weight: 400;
+    font-family: "Montserrat", sans-serif;
+    font-weight: 600;
   }
   
   .background {
