+++ src/pages/Home/Home.module.css
@@ -1,0 +1,29 @@
+.container {
+  width: 100%;
+  height: 20%;
+}
+.card{
+  width: 400px;
+  height: 200px;
+  background-color: #222831;
+  color: #fff;
+  border-radius: 20px;
+}
+.dados{
+  margin-top: 15px;
+  margin-left: 30px;
+}
+.projeto{
+  margin: 10px;
+}
+.tarefas{
+  margin: 10px;
+}
+.titulo{
+  text-align: center;
+  padding: 5px;
+}
+.descricao{
+  padding: 5px;
+  text-align: justify;
+}
