--- src/pages/Home/ListaDeProjetos.jsx
@@ -1,15 +1,0 @@
-import Home from "./Home";
-import React, {useState} from "react";
-
-
-export default function ListaDeProjetos() {
-  const [projeto, setProjeto] = useState({
-    tituloProjeto: "TaskSync",
-    descricaoProjeto: "AAAAAAAAAAAAAA",
-  });
-  return (
-    <div>
-      <Home projeto={projeto} />
-    </div>
-  );
-}
