+++ src/pages/HomeOld/Home.css
@@ -1,0 +1,122 @@
+body{
+  font-family: "Montserrat", sans-serif;
+  font-weight: 900;
+}
+.container {
+  position: relative;
+}
+
+.div-top {
+  height: 75px;
+  background-color: rgb(255, 255, 255);
+  text-align: right;
+  box-shadow: rgba(0, 0, 0, 0.15) 2.4px 2.4px 3.2px;
+  position: relative;
+}
+.div-left {
+  align-items: center;
+  background-color: #d9d9d9;
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 100vh;
+  width: 20%;
+  z-index: 2;
+  display: flex;
+  flex-direction: column; 
+  justify-content: center;
+  box-shadow: rgba(0, 0, 0, 0.15) 2.4px 2.4px 3.2px;
+}
+.novos {
+  height: 50px;
+  width: 135px;
+  border-radius: 15px;
+  background-color: #222831;
+  color: white;
+  margin: 10px;
+}
+.card {
+  background-color: #222831;
+  color: white;
+  margin-left: 100px;
+  border: 1px solid;
+  padding: 10px;
+  height: 115px;
+  width: 290px;
+  border-radius: 15px;
+  margin: 15px;
+}
+.title {
+  text-align: center;
+  padding: 5px;
+}
+.descricao{
+  padding: 5px;
+  text-align: justify;
+}
+.projetos {
+  padding: 10px;
+  margin: 10px;
+}
+.title-project{
+  padding-right: 5px;
+  padding-top: 30px;
+  text-align: right;
+}
+.div-bottom {
+  margin: auto;
+  width: 800px;
+  height: 630px;
+}
+
+.botao-menu{
+  margin: 30px;
+  background-image: url("/src/assets/HomePage.png");
+  background-size: 40px 40px;
+  background-position-y: 10px;
+  background-position-x: 20px;
+  background-repeat: no-repeat;
+  width: 80px;
+  height: 60px;
+  border-radius: 20px;
+  background-color: #222831;
+
+}
+.botao-projetos{
+  margin: 30px;
+  background-image: url("/src/assets/Settings.png");
+  background-size: 40px 40px;
+  background-position-y: 10px;
+  background-position-x: 20px;
+  background-repeat: no-repeat;
+  width: 80px;
+  height: 60px;
+  border-radius: 20px;
+  background-color: #222831;
+
+}
+.botao-calendario{
+  margin: 30px;
+  background-image: url("/src/assets/Calendar31.png");
+  background-size: 40px 40px;
+  background-position-y: 10px;
+  background-position-x: 20px;
+  background-repeat: no-repeat;
+  width: 80px;
+  height: 60px;
+  border-radius: 20px;
+  background-color: #222831;
+  
+}
+.botao-config {  
+  margin: 30px;
+  background-image: url("/src/assets/Settings.png");
+  background-size: 40px 40px;
+  background-position-y: 10px;
+  background-position-x: 20px;
+  background-repeat: no-repeat;
+  width: 80px;
+  height: 60px;
+  border-radius: 20px;
+  background-color: #222831;
+}
