+++ src/pages/HomeOld/ListaDeProjetos.jsx
@@ -1,0 +1,6 @@
+import Home from "./Home";
+import React, { useState } from "react";
+
+export default function ListaDeProjetos() {
+  return <div></div>;
+}
